requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
pymupdf
python-dotenv
streamlit
supabase
